TD-Jour6/app/data_api/requirements.txt
flask==3.0.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
prometheus-client==0.19.0
